+++ server/config.yaml
@@ -1,19 +1,19 @@
-port: 7003 ## webserver port, which you can access in web
+port: 7001 ## webserver port, which you can access in web
 gateway: 
-  ip: hostIP
+  ip: 192.168.8.44
   port: 8090
   https: false
   runmode: dev
 stats-exporter:
-  ip: hostIP
+  ip: 192.168.8.44
   port: 9200
   https: false
 node-exporter:
-  - ip: hostIP
+  - ip: 192.168.8.143
     port: 9100
     https: false
 prometheus: 
-  ip: hostIP
+  ip: 192.168.8.44
   port: 9090
   https: false
   scrape_interval: 5s
@@ -23,17 +23,17 @@ nebula-cluster:
   name: 'default' # cluster name
   metad: 
     - name: metad0
-      endpointIP: nebulaMetadIP
+      endpointIP: 192.168.8.143
       port: 9559
       endpointPort: 19559
   graphd:
     - name: graphd0
-      endpointIP: nebulaGraphdIP
+      endpointIP: 192.168.8.143
       port: 9669
       endpointPort: 19669
   storaged:
     - name: storaged0
-      endpointIP: nebulaStoragedIP
+      endpointIP: 192.168.8.143
       port: 9779
       endpointPort: 19779
